working login and reg, doesnot rerout yet

+++ server/routes/register.js
@@ -1,4 +1,5 @@
 var express = require("express");
+const { doesNotMatch } = require("node:assert");
 var router = express.Router();
 const fs = require("node:fs");
 const path = require("path");
@@ -7,30 +8,36 @@ const path = require("path");
 const PathToDb = path.join(__dirname, "../db/users.json");
 
 router.post("/", (req, res) => {
-    console.log("req.body", req.body);
-    let username = req.body.username;
-    let password = req.body.password;
-    let user = {
-        username: username,
-        password: password,
-    };
+  console.log("req.body", req.body);
+  let username = req.body.username;
+  let password = req.body.password;
+  let user = {
+    username: username,
+    password: password,
+  };
 
-    // let check1 = firstChecks(username, password)
-    if (typeof username === "string" && username.length > 0 && password.length > 0) {
-        console.log("first check passed");
-        let usersjson = fs.readFileSync(PathToDb, "utf-8");
-        console.log("usersjson: ", usersjson);
-        let users = JSON.parse(usersjson);
-        let newId = users[users.length - 1].id + 1;
-        console.log(users[users.length - 1].id);
-        user.id = newId;
-        users.push(user);
-        usersjson = JSON.stringify(users);
-        fs.writeFileSync(PathToDb, usersjson);
-        console.log("i passed the commiting code");
-    }
+  // let check1 = firstChecks(username, password)
+  if (typeof username === "string" && username.length > 0 && password.length > 0) {
+    let usersjson = fs.readFileSync(PathToDb, "utf-8");
+    let users = JSON.parse(usersjson);
+    doesUserAlreadyExists = users.filter((u) => username === u.username && password === u.password);
+    if (doesUserAlreadyExists.length !== 0)
+      return res.status(401).json({
+        message: "user already exists",
+      });
 
-    res.send(JSON.stringify(user));
+    let newId = users[users.length - 1].id + 1;
+    user.id = newId;
+    users.push(user);
+    usersjson = JSON.stringify(users);
+    fs.writeFileSync(PathToDb, usersjson);
+  }
+
+  res.send(
+    JSON.stringify({
+      username: username,
+    })
+  );
 });
 
 module.exports = router;
